Base project

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -4,7 +4,10 @@ namespace App\Http\Controllers\Admin;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Admin\BaseController;
+use App\Models\Role;
 use Spatie\Permission\Models\Permission;
+use Illuminate\Support\Facades\DB;
+
 
 class PermissionController extends BaseController
 {
@@ -18,7 +21,7 @@ class PermissionController extends BaseController
     {
         $permissions = Permission::get();
         $head = $this->getHead();
-        $this->response['head'] = array_merge(['name', 'guard_name', 'created_at'], $head);
+        $this->response['head'] = array_merge([trans('admin.name'), trans('admin.guard_name'), trans('admin.created_at')], $head);
         $this->response['permissions'] = $permissions;
 
         return view('admin.role-permission.permission.view', $this->response);
@@ -26,48 +29,39 @@ class PermissionController extends BaseController
 
     public function create()
     {
+        $this->response['guards'] = Role::getDefaultGuards();
         return view('admin.role-permission.permission.create', $this->response);
     }
 
     public function store(Request $request)
     {
         $request->validate([
-            'name' => [
-                'required',
-                'string',
-                'unique:permissions,name'
-            ],
-            'guard_name' => [
-                'required',
-                'string'            ]
+            'name' => 'required|regex:/^[a-zA-Z0-9_\ ]+$/|unique:permissions,name|min:4|max:16',
+            'guard_name' => 'required|string'
         ]);
 
-        Permission::create([
+        $permission = Permission::create([
             'name' => $request->name,
             'guard_name' => $request->guard_name,
         ]);
+        
+        $this->clearCache();
 
-        return redirect(route('admin.permission.index'))->with('status','Permission Created Successfully');
+        return redirect(route('admin.permission.edit', [$permission->id]))->with('status','Permission Created Successfully');
     }
 
     public function edit(Permission $permission)
     {
         $this->response['permission'] = $permission;
+        $this->response['guards'] = Role::getDefaultGuards();
         return view('admin.role-permission.permission.update', $this->response);
     }
 
     public function update(Request $request, Permission $permission)
     {
         $request->validate([
-            'name' => [
-                'required',
-                'string',
-                'unique:permissions,name,'.$permission->id
-            ],
-            'guard_name' => [
-                'required',
-                'string'
-            ]
+            'name' => 'required|regex:/^[a-zA-Z0-9_\ ]+$/|unique:permissions,name,' . $permission->id . '|min:4|max:16',
+            'guard_name' => 'required|string'
         ]);
 
         $permission->update([
@@ -75,13 +69,21 @@ class PermissionController extends BaseController
             'guard_name' => $request->guard_name,
         ]);
 
+        $this->clearCache();
+
         return redirect(route('admin.permission.edit', [$permission->id]))->with('status','Permission Updated Successfully');
     }
 
     public function destroy($permissionId)
     {
-        $permission = Permission::find($permissionId);
-        $permission->delete();
-        return redirect(route('admin.permission.index'))->with('status','Permission Deleted Successfully');
+        try {
+            $permission = Permission::findOrFail($permissionId);
+            if ($permission) {
+                $permission->delete();
+                return redirect(route('admin.permission.index'))->with('status','Permission Deleted Successfully');        
+            }
+        } catch (\Exception $e) {
+            return redirect(route('admin.permission.index'))->with('status','Permission Couldn\'t be deleted Successfully');
+        }
     }
 }
